Add opsCollectionPerDoc option to Mongo storage.
If this is true, each document will have its own collection for its operations. If false, operations are put in a single large operations table keyed by docname. The former is faster and requires less storage, but does not easily scale beyond 12,000 or so documents (the maximum number of collections per database) and does not seem compatible with Mongo's sharding.